all config

+++ src/main/java/com/example/Venus/repo/BannerRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface BannerRepo extends JpaRepository<Banners, Integer> {
+public interface BannerRepo extends JpaRepository<Banners, Long> {
     Optional<Banners> findByUserIdAndTitle(Long userId, String title);
 }
