+++ src/main/java/com/example/Venus/repo/EnquiryRepo.java
@@ -3,6 +3,8 @@ package com.example.Venus.repo;
 import com.example.Venus.entities.Enquiry;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnquiryRepo extends JpaRepository<Enquiry, Integer> {
+import java.util.List;
+
+public interface EnquiryRepo extends JpaRepository<Enquiry, Long> {
 
 }
